Update ProcessTemplateJob.php
Fix npm

+++ app/Jobs/ProcessTemplateJob.php
@@ -172,8 +172,10 @@ final class ProcessTemplateJob implements ShouldQueue
 
         $this->commands[] = 'ddev composer run post-root-package-install';
         $this->commands[] = 'ddev php artisan key:generate --ansi';
-
-        $this->commands[] = 'docker run --rm -v ${PWD}:/app python:3-alpine python /app/.ddev/bun/package_json_updater.py';
+        
+        if ($config['javascript_package_manager'] === JavascriptPackageManagerEnum::Bun->value) {
+                $this->commands[] = 'docker run --rm -v ${PWD}:/app python:3-alpine python /app/.ddev/bun/package_json_updater.py';
+            }
 
         $this->installNodeDependencies();
